feat(core): make possessive nouns imply compound nouns
