+++ harper-core/dictionary.dict
@@ -48479,7 +48479,7 @@ waxiness/1M
 waxwing/1SM
 waxwork/1SM
 waxy/51RPT
-way/145SM~
+way/15SM~
 waybill/14SM
 wayfarer/1MS
 wayfaring/514SM
